+++ apps/backend/src/modules/project-groups/project-group.entity.ts
@@ -6,25 +6,31 @@ import { type ProjectGroupCreateResponseDto } from "./libs/types/types.js";
 import { type ProjectPermissionModel } from "./project-permission.model.js";
 
 class ProjectGroupEntity implements Entity {
+	private createdAt: null | string;
 	private id: null | number;
 	private name!: string;
-	private permissions!: Pick<ProjectPermissionModel, "id">[];
+	private permissions!: Partial<Pick<ProjectPermissionModel, "id" | "name">>[];
 	private projectId!: Pick<ProjectModel, "id">;
-	private users!: Pick<UserModel, "id">[];
+	private users!: Partial<
+		Pick<UserModel, "createdAt" | "email" | "id" | "name">
+	>[];
 
 	private constructor({
+		createdAt,
 		id,
 		name,
 		permissions,
 		projectId,
 		users,
 	}: {
+		createdAt: null | string;
 		id: null | number;
 		name: string;
-		permissions: Pick<ProjectPermissionModel, "id">[];
+		permissions: Partial<Pick<ProjectPermissionModel, "id" | "name">>[];
 		projectId: Pick<ProjectModel, "id">;
-		users: Pick<UserModel, "id">[];
+		users: Partial<Pick<UserModel, "createdAt" | "email" | "id" | "name">>[];
 	}) {
+		this.createdAt = createdAt;
 		this.id = id;
 		this.name = name;
 		this.permissions = permissions;
@@ -33,23 +39,26 @@ class ProjectGroupEntity implements Entity {
 	}
 
 	public static initialize({
+		createdAt,
 		id,
 		name,
 		permissions,
-		projectId,
+		projectId: project,
 		users,
 	}: {
+		createdAt: string;
 		id: number;
 		name: string;
-		permissions: Pick<ProjectPermissionModel, "id">[];
+		permissions: Pick<ProjectPermissionModel, "id" | "name">[];
 		projectId: Pick<ProjectModel, "id">;
-		users: Pick<UserModel, "id">[];
+		users: Pick<UserModel, "createdAt" | "email" | "id" | "name">[];
 	}): ProjectGroupEntity {
 		return new ProjectGroupEntity({
+			createdAt,
 			id,
 			name,
 			permissions,
-			projectId,
+			projectId: project,
 			users,
 		});
 	}
@@ -66,6 +75,7 @@ class ProjectGroupEntity implements Entity {
 		users: Pick<UserModel, "id">[];
 	}): ProjectGroupEntity {
 		return new ProjectGroupEntity({
+			createdAt: null,
 			id: null,
 			name,
 			permissions,
@@ -82,19 +92,34 @@ class ProjectGroupEntity implements Entity {
 	} {
 		return {
 			name: this.name,
-			permissions: this.permissions,
+			permissions: this.permissions as Pick<
+				ProjectPermissionModel,
+				"id" | "name"
+			>[],
 			projectId: this.projectId,
-			users: this.users,
+			users: this.users as Pick<
+				UserModel,
+				"createdAt" | "email" | "id" | "name"
+			>[],
 		};
 	}
 
 	public toObject(): ProjectGroupCreateResponseDto {
 		return {
+			createdAt: this.createdAt as string,
 			id: this.id as number,
 			name: this.name,
-			permissions: this.permissions,
+			permissions: this.permissions.map((permission) => ({
+				id: permission.id,
+				name: permission.name,
+			})) as Pick<ProjectPermissionModel, "id" | "name">[],
 			projectId: this.projectId,
-			users: this.users,
+			users: this.users.map((user) => ({
+				createdAt: user.createdAt,
+				email: user.email,
+				id: user.id,
+				name: user.name,
+			})) as Pick<UserModel, "createdAt" | "email" | "id" | "name">[],
 		};
 	}
 }