+++ apps/backend/src/modules/project-groups/project-group.repository.ts
@@ -1,8 +1,13 @@
 import { transaction } from "objection";
 
+import { SortType } from "~/libs/enums/enums.js";
 import { changeCase } from "~/libs/helpers/helpers.js";
 import { HTTPCode } from "~/libs/modules/http/libs/enums/enums.js";
-import { type Repository } from "~/libs/types/types.js";
+import {
+	type PaginationQueryParameters,
+	type PaginationResponseDto,
+	type Repository,
+} from "~/libs/types/types.js";
 
 import { ExceptionMessage } from "./libs/enums/enums.js";
 import { ProjectGroupError } from "./libs/exceptions/exceptions.js";
@@ -27,7 +32,7 @@ class ProjectGroupRepository implements Repository {
 				key,
 				name,
 				permissions,
-				projects: projectId,
+				projects: [projectId],
 				users,
 			};
 
@@ -35,16 +40,13 @@ class ProjectGroupRepository implements Repository {
 				.query(trx)
 				.insertGraph(projectGroupData, { relate: true })
 				.returning("*")
-				.withGraphJoined("[permissions, projects, users]");
+				.withGraphFetched("[permissions, projects, users]");
 
 			await trx.commit();
 
 			return ProjectGroupEntity.initialize({
-				id: createdProjectGroup.id,
-				name: createdProjectGroup.name,
-				permissions,
+				...createdProjectGroup,
 				projectId,
-				users,
 			});
 		} catch {
 			await trx.rollback();
@@ -66,6 +68,29 @@ class ProjectGroupRepository implements Repository {
 
 	public findAll(): ReturnType<Repository["findAll"]> {
 		return Promise.resolve({ items: [] });
+	}
+
+	public async findAllByProjectId(
+		id: number,
+		{ page, pageSize }: PaginationQueryParameters,
+	): Promise<PaginationResponseDto<ProjectGroupEntity>> {
+		const { results, total } = await this.projectGroupModel
+			.query()
+			.orderBy("createdAt", SortType.DESCENDING)
+			.page(page, pageSize)
+			.joinRelated("projects")
+			.where("projects.id", id)
+			.withGraphFetched("[permissions, users, projects]");
+
+		return {
+			items: results.map((projectGroup) =>
+				ProjectGroupEntity.initialize({
+					...projectGroup,
+					projectId: { id },
+				}),
+			),
+			totalItems: total,
+		};
 	}
 
 	public async findByName(name: string): Promise<null | ProjectGroupModel> {