+++ apps/frontend/src/index.tsx
@@ -3,6 +3,7 @@ import { createRoot } from "react-dom/client";
 
 import {
 	App,
+	Navigate,
 	ProtectedRoute,
 	RouterProvider,
 	StoreProvider,
@@ -13,11 +14,13 @@ import { store } from "~/libs/modules/store/store.js";
 import { AccessManagement } from "~/pages/access-management/access-management.jsx";
 import { Analytics } from "~/pages/analytics/analytics.jsx";
 import { Auth } from "~/pages/auth/auth.jsx";
-import { Home } from "~/pages/home/home.jsx";
+import { NoAccess } from "~/pages/no-access/no-access.jsx";
 import { NotFound } from "~/pages/not-found/not-found.jsx";
 import { Profile } from "~/pages/profile/profile.jsx";
 import { Project } from "~/pages/project/project.jsx";
 import { Projects } from "~/pages/projects/projects.jsx";
+
+import { ProjectAccessManagement } from "./pages/project-access-management/project-access-management.js";
 
 createRoot(document.querySelector("#root") as HTMLElement).render(
 	<StrictMode>
@@ -65,9 +68,13 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 							{
 								element: (
 									<ProtectedRoute>
-										<Home />
+										<NoAccess />
 									</ProtectedRoute>
 								),
+								path: AppRoute.NO_ACCESS,
+							},
+							{
+								element: <Navigate to={AppRoute.PROJECTS} />,
 								path: AppRoute.ROOT,
 							},
 							{
@@ -81,6 +88,10 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 							{
 								element: <Project />,
 								path: AppRoute.PROJECT,
+							},
+							{
+								element: <ProjectAccessManagement />,
+								path: AppRoute.PROJECT_ACCESS_MANAGEMENT,
 							},
 						],
 						element: <App />,