+++ apps/frontend/src/libs/components/table/table.tsx
@@ -4,8 +4,10 @@ import {
 	useReactTable,
 } from "@tanstack/react-table";
 
+import { Loader } from "~/libs/components/components.js";
 import { EMPTY_LENGTH } from "~/libs/constants/constants.js";
 import { getValidClassNames } from "~/libs/helpers/helpers.js";
+import { useEffect, useRef, useState } from "~/libs/hooks/hooks.js";
 import { type TableColumn } from "~/libs/types/types.js";
 
 import { SelectRowCell } from "./libs/components/components.js";
@@ -15,6 +17,7 @@ type BaseProperties<T> = {
 	columns: TableColumn<T>[];
 	data: T[];
 	emptyPlaceholder?: string;
+	isLoading?: boolean;
 	isScrollDisabled?: boolean;
 };
 
@@ -32,6 +35,7 @@ const Table = <T extends object>({
 	columns,
 	data,
 	emptyPlaceholder = "There is nothing yet.",
+	isLoading,
 	isScrollDisabled,
 	...selectableProperties
 }: Properties<T>): JSX.Element => {
@@ -45,12 +49,26 @@ const Table = <T extends object>({
 		getCoreRowModel: getCoreRowModel(),
 	});
 
+	const tbodyReference = useRef<HTMLTableSectionElement>(null);
+	const [tableBodyHeight, setTableBodyHeight] = useState<null | number>(null);
+
+	useEffect(() => {
+		if (!isLoading && tbodyReference.current) {
+			setTableBodyHeight(tbodyReference.current.clientHeight);
+		}
+	}, [isLoading]);
+
 	const hasData = data.length !== EMPTY_LENGTH;
 	const isRowSelectable = typeof onRowSelect === "function";
 
 	return (
 		<div className={styles["table-container"]}>
-			<table className={styles["table"]}>
+			<table
+				className={getValidClassNames(
+					styles["table"],
+					isScrollDisabled && styles["table-no-scroll"],
+				)}
+			>
 				<thead className={styles["table-head"]}>
 					{table.getHeaderGroups().map((headerGroup) => (
 						<tr className={styles["table-row"]} key={headerGroup.id}>
@@ -77,13 +95,20 @@ const Table = <T extends object>({
 						</tr>
 					))}
 				</thead>
-				<tbody
-					className={getValidClassNames(
-						styles["table-body"],
-						isScrollDisabled && styles["table-body-no-scroll"],
+				<tbody className={styles["table-body"]} ref={tbodyReference}>
+					{isLoading && (
+						<tr className={styles["table-row"]}>
+							<td
+								className={styles["table-loader"]}
+								colSpan={columns.length}
+								style={{ height: tableBodyHeight || "auto" }}
+							>
+								<Loader />
+							</td>
+						</tr>
 					)}
-				>
-					{hasData ? (
+					{!isLoading &&
+						hasData &&
 						table.getRowModel().rows.map((row) => (
 							<tr className={styles["table-row"]} key={row.id}>
 								{isRowSelectable && (
@@ -112,8 +137,9 @@ const Table = <T extends object>({
 									</td>
 								))}
 							</tr>
-						))
-					) : (
+						))}
+
+					{!isLoading && !hasData && (
 						<tr className={styles["table-row"]}>
 							<td className={styles["table-data"]} colSpan={columns.length}>
 								<p className={styles["empty-placeholder"]}>