+++ apps/frontend/src/libs/enums/notification-message.enum.ts
@@ -7,6 +7,7 @@ const NotificationMessage = {
 	PROJECT_API_KEY_CREATE_SUCCESS: "Project API key was successfully generated.",
 	PROJECT_CREATE_SUCCESS: "Project was successfully created.",
 	PROJECT_DELETE_SUCCESS: "Project was successfully deleted.",
+	PROJECT_GROUP_CREATE_SUCCESS: "Project group was successfully created.",
 	PROJECT_UPDATE_SUCCESS: "Project was successfully updated.",
 	USER_DELETE_SUCCESS: "User deleted successfully.",
 } as const;