+++ apps/frontend/src/pages/access-management/libs/helpers/get-group-columns.helper.tsx
@@ -13,23 +13,23 @@ const getGroupColumns = (actions: {
 	{
 		accessorKey: "name",
 		header: "Name",
-		size: 100,
+		size: 200,
 	},
 	{
 		accessorFn: (group: GroupRow): string => group.permissions.join(", "),
 		header: "Permissions",
-		size: 200,
+		size: 250,
 	},
 	{
 		accessorKey: "userCount",
 		header: "Users",
-		size: 50,
+		size: 100,
 	},
 	{
 		accessorFn: (group: GroupRow): string =>
 			formatDate(new Date(group.createdAt), "d MMM yyyy HH:mm"),
 		header: "Created At",
-		size: 100,
+		size: 200,
 	},
 	{
 		cell: ({ row: { original: group } }) => (
@@ -43,7 +43,7 @@ const getGroupColumns = (actions: {
 		),
 		header: "",
 		id: "menu",
-		size: 0,
+		size: 60,
 	},
 ];