+++ apps/frontend/src/pages/project-access-management/libs/components/project-group-create-form/libs/constants/default-project-group-create-payload.constant.ts
@@ -1,0 +1,10 @@
+import { type ProjectGroupCreateRequestDto } from "~/modules/project-groups/project-groups.js";
+
+const DEFAULT_PROJECT_GROUP_CREATE_PAYLOAD: ProjectGroupCreateRequestDto = {
+	name: "",
+	permissionIds: [],
+	projectId: 0,
+	userIds: [],
+};
+
+export { DEFAULT_PROJECT_GROUP_CREATE_PAYLOAD };