+++ packages/shared/src/modules/projects/libs/types/project-get-all-request-dto.type.ts
@@ -1,5 +1,7 @@
+import { type PaginationQueryParameters } from "../../../../libs/types/types.js";
+
 type ProjectGetAllRequestDto = {
 	name: string;
-};
+} & PaginationQueryParameters;
 
 export { type ProjectGetAllRequestDto };