+++ scripts/analytics/src/libs/enums/enums.ts
@@ -1,0 +1,6 @@
+export {
+	APIPath,
+	AppEnvironment,
+	ContentType,
+	ServerErrorType,
+} from "@git-fit/shared";