+++ scripts/analytics/src/libs/modules/config/libs/types/environment-schema.type.ts
@@ -1,0 +1,8 @@
+type EnvironmentSchema = {
+	APP: {
+		API_ORIGIN_URL: string;
+		SERVER_URL: string;
+	};
+};
+
+export { type EnvironmentSchema };