+++ scripts/analytics/src/modules/analytics/libs/constants/constants.ts
@@ -1,0 +1,1 @@
+export { COMMIT_REGEX } from "./commit-regex.constant.js";